splitting train_utils file
