+++ misc/pytorch_toolkit/chest_xray_screening_federated_gcn/src/utils/train_utils_gnn.py
@@ -3,6 +3,7 @@ from .train_utils import train_end_to_end, initialize_training
 from .inference_utils import inference
 from .train_utils import initialize_model_weights, instantiate_architecture
 from .misc import aggregate_local_weights, save_model_weights
+from .transformations import train_transform, test_transform
 import copy
 
 def lcl_train_gnn(lr, trn_loader, val_loader, criterion, cnv_lyr,
@@ -44,9 +45,7 @@ def trainer_with_GNN(lr, b_sz, img_pth, split_npz, train_transform, test_transfo
     backbone_model = backbone_model.to(device)
     gnn_model = gnn_model.to(device)
 
-   
-    ############## Initialize Data Loaders #################
-    
+    ############## Initialize Data Loaders #################    
     trn_loader0, val_loader0, criterion0, edge_index0, edge_attr0=initialize_training(0, img_pth, split_npz, 
                                                               train_transform, test_transform, b_sz, device=device)
     
@@ -62,11 +61,9 @@ def trainer_with_GNN(lr, b_sz, img_pth, split_npz, train_transform, test_transfo
     trn_loader4, val_loader4, criterion4, edge_index4, edge_attr4=initialize_training(4, img_pth, split_npz, 
                                                               train_transform, test_transform, b_sz, device=device)
     
-    
     ### Initialize local and global model weights with the Imagenet pre-trained weights for backbone 
     #and identical model weights for the other layers.
 
-    
     glbl_cnv_wt, glbl_backbone_wt, glbl_fc_wt, gnn_wt=initialize_model_weights(cnv_lyr, backbone_model, 
                                                                                     fc_layers, gnn_model)
     sit0_gnn_wt=copy.deepcopy(gnn_wt)
@@ -88,12 +85,10 @@ def trainer_with_GNN(lr, b_sz, img_pth, split_npz, train_transform, test_transfo
         sit3_gnn_wt=checkpoint['sit3_gnn_model']
         sit4_gnn_wt=checkpoint['sit4_gnn_model']
     
-    
     ################ Begin Actual Training ############
     max_val=0
     for epoch in range(0, max_epochs):
         print('############ Epoch: '+str(epoch)+'   #################')
-        
         ###### Load the global model weights ########
         cnv_lyr.load_state_dict(glbl_cnv_wt)
         backbone_model.load_state_dict(glbl_backbone_wt)
@@ -145,7 +140,6 @@ def trainer_with_GNN(lr, b_sz, img_pth, split_npz, train_transform, test_transfo
                                                 criterion4, cnv_lyr, backbone_model,fc_layers, gnn_model,
                                                                                  edge_index4, edge_attr4, device)
 
-        
         avg_val=(prv_val0+prv_val1+prv_val2+prv_val3+prv_val4)/5
         print('Avg Val AUC: '+str(avg_val))
         
@@ -157,7 +151,6 @@ def trainer_with_GNN(lr, b_sz, img_pth, split_npz, train_transform, test_transfo
                                sit2_gnn_wt, sit3_gnn_wt, sit4_gnn_wt)
             print('Validation Performance Improved !')
             
-            
         ############### Compute the global model weights #############
         
         glbl_cnv_wt=aggregate_local_weights(sit0_cnv_wt, sit1_cnv_wt, sit2_cnv_wt,
@@ -167,3 +160,13 @@ def trainer_with_GNN(lr, b_sz, img_pth, split_npz, train_transform, test_transfo
                                                 sit3_backbone_wt, sit4_backbone_wt, device)
         
         glbl_fc_wt=aggregate_local_weights(sit0_fc_wt, sit1_fc_wt, sit2_fc_wt, sit3_fc_wt, sit4_fc_wt, device)
+
+def train_model(config):
+    if torch.cuda.is_available() and config['gpu']=='True':
+        device = torch.device('cuda')
+    else:
+        device = torch.device('cpu')
+    trainer_with_GNN(config['lr'],config['batch_size'], config['data'], config['split_npz'],
+                        train_transform, test_transform, config['epochs'], config['backbone'],
+                        device, config['checkpoint'], config['savepath'] )
+
